Refactor CEL
