+++ pkg/utils/cel/doc.go
@@ -1,5 +1,5 @@
 /*
-Copyright 2023 The Kubernetes Authors.
+Copyright 2024 The Kubernetes Authors.
 
 Licensed under the Apache License, Version 2.0 (the "License");
 you may not use this file except in compliance with the License.
@@ -14,5 +14,5 @@ See the License for the specific language governing permissions and
 limitations under the License.
 */
 
-// Package cel implements a metrics evaluator using the Common Expression Language (CEL).
+// Package cel implements a environment for the Common Expression Language (CEL).
 package cel
