+++ pkg/utils/cel/resource_list.go
